Fix unit tests during merge

+++ internal/transform/schema.go
@@ -115,22 +115,22 @@ type LiquidityPoolAsset struct {
 
 // PoolOutput is a representation of a liquidity pool that aligns with the Bigquery table liquidity_pools
 type PoolOutput struct {
-	PoolID             string `json:"liquidity_pool_id"`
-	PoolType           string `json:"type"`
-	PoolFee            uint32 `json:"fee"`
-	TrustlineCount     uint64 `json:"trustline_count"`
-	PoolShareCount     uint64 `json:"pool_share_count"`
-	AssetAType         string `json:"asset_a_type"`
-	AssetACode         string `json:"asset_a_code"`
-	AssetAIssuer       string `json:"asset_a_issuer"`
-	AssetAReserve      uint64 `json:"asset_a_amount"`
-	AssetBType         string `json:"asset_b_type"`
-	AssetBCode         string `json:"asset_b_code"`
-	AssetBIssuer       string `json:"asset_b_issuer"`
-	AssetBReserve      uint64 `json:"asset_b_amount"`
-	LastModifiedLedger uint32 `json:"last_modified_ledger"`
-	LedgerEntryChange  uint32 `json:"ledger_entry_change"`
-	Deleted            bool   `json:"deleted"`
+	PoolID             string  `json:"liquidity_pool_id"`
+	PoolType           string  `json:"type"`
+	PoolFee            uint32  `json:"fee"`
+	TrustlineCount     uint64  `json:"trustline_count"`
+	PoolShareCount     float64 `json:"pool_share_count"`
+	AssetAType         string  `json:"asset_a_type"`
+	AssetACode         string  `json:"asset_a_code"`
+	AssetAIssuer       string  `json:"asset_a_issuer"`
+	AssetAReserve      float64 `json:"asset_a_amount"`
+	AssetBType         string  `json:"asset_b_type"`
+	AssetBCode         string  `json:"asset_b_code"`
+	AssetBIssuer       string  `json:"asset_b_issuer"`
+	AssetBReserve      float64 `json:"asset_b_amount"`
+	LastModifiedLedger uint32  `json:"last_modified_ledger"`
+	LedgerEntryChange  uint32  `json:"ledger_entry_change"`
+	Deleted            bool    `json:"deleted"`
 }
 
 // AssetOutput is a representation of an asset that aligns with the BigQuery table history_assets
